requirements.txt
tencentcloud-sdk-python-common==3.0.1423
tencentcloud-sdk-python-ssl==3.0.1423